Fitting masked pixels
Implemented lines so that the code only fits flux pixels for which mask ==1. the rest are set to zero. Also fixed the file paths because for open() the code starts from working dir and not from module directory. whereas for saving you are starting from the module directory  (I think)