Hot/dead pixel detection
Implemented a function in the preprocess module to automatically detect hot/dead pixels and set them to 0, + setting their error to 1e6 so that they do not impact the fit. Also fixed some of the if loops in the pre-processing to allow for smooth fitting of FRESCO GOODS-S data (bc here the headers are a bit different which is annoying...)
Will have to somehow homogenize this later on.