Added sampling for PA orientation and EL wavelength
In the one_comp model, I am now sampling whether a +180 has to be added to the PA. Also sampling for slight shifits in the EL centroid with a very narrow gaussian prior.
Added the XLA command at the top and removed plt.imshows to see if this speeds to code up on GPUs at all.