Fully integrated models, grism, and utils modules
Made all of the ncessary modifications so that now those modules are fully integrated and any redundundancies are removed.
For now, I only implemented the 1 component model which is now called Disk, but wull implement all of the others as I go.
Also fixed an error: the grism dispersion has to be done in low resolution because there is no way to spatially ressample the x axis of the grism.
==> so fluxes are sampled in high res (and vels computed in high res) but then everything is scaled down pre-dispersion.
Will probably remove the factor variable since I will never use it.