Corrected dispersion function
I was dispersing the pixels wrong, but it is done properly now. I was only dispersing the centeral pixel and then spreading the rest around according the their delta x in the image frame but each pixel needs to be dispersed itself. THis makes the EL map more squished than the image which makes sense.

Also added a flux scaling variable that I sample in the Disk model