fixed fengwu's lsf
the fwhm was is nm which was really throwing it off, but it' still not fully correct. For now, I manually corrected it wiht a 0.617 factor for the sigma0 but it doesn't match at all wavelengths. need to look into this
also fixed the path to the nircam_grism config files