create run_fit function
instead of using the main part of fitting to send off the fitting, I made a function inside the module and moved the actual interface in another RunGeko.py pything script that is in grism_project. That way the package itself doesn't need to be called to run.. which doesn't make sense anyways!